% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jb_eq5dGetValues.R
\name{jb_eq5dGetValues}
\alias{jb_eq5dGetValues}
\title{jb_eq5dGetValues: Get summary stats from EQ-5D
 
 @author James Black
 @param barorder Numeric 1 to whatever. For which bar (or timepoint)  this set of EQ-5D measurements will be on the plot.
 @param eq_index The main EQ-5D index score. 
 @param domain_one Any of the five eq domains - order doesn't matter
 @param domain_two Any of the five eq domains - order doesn't matter
 @param domain_three Any of the five eq domains - order doesn't matter
 @param domain_four Any of the five eq domains - order doesn't matter
 @param domain_five Any of the five eq domains - order doesn't matter
 @examples
 "See epijim.uk"}
\usage{
jb_eq5dGetValues(barorder, eq_index, domain_one, domain_two, domain_three,
  domain_four, domain_five)
}
\description{
jb_eq5dGetValues: Get summary stats from EQ-5D
 
 @author James Black
 @param barorder Numeric 1 to whatever. For which bar (or timepoint)  this set of EQ-5D measurements will be on the plot.
 @param eq_index The main EQ-5D index score. 
 @param domain_one Any of the five eq domains - order doesn't matter
 @param domain_two Any of the five eq domains - order doesn't matter
 @param domain_three Any of the five eq domains - order doesn't matter
 @param domain_four Any of the five eq domains - order doesn't matter
 @param domain_five Any of the five eq domains - order doesn't matter
 @examples
 "See epijim.uk"
}

