% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jb_Load.R
\name{jb_load}
\alias{jb_load}
\title{Automatically install missing packages}
\usage{
jb_load(package)
}
\arguments{
\item{name}{The name of the package.}
}
\description{
Automatically install missing packages
}
\examples{
jb_load()
}
\keyword{packages}

