% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jb_summaryfunctions.R
\name{jb_multihists}
\alias{jb_multihists}
\title{Quickly histograph a bunch of variables}
\usage{
jb_multihists(x)
}
\arguments{
\item{data}{Dataframe, will automatically filter to numeric. But limited to max that fit on one page.}
}
\description{
Packager loader that will automatically install missing packages
}
\examples{
data <- iris
jb_multihists(data)

}
\keyword{summary}

