% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jb_eq5d.R
\name{jb_ploteq5d}
\alias{jb_ploteq5d}
\title{bulleteq5d: Vertical Bullet Graph
 
 @author James Black - a ggplot take on the bullet plots by Simon Müller
 
 ""              ""         "" ""            "timepoints" 
 @param data:  data.frame that is produced by rbinding runs of jb_geteq5dvalues
                    bar  : number of bar values are from
                    eq_index     : eq5d index score
                    impaired_domains     : number of domains listed as impaired
                    
 @param bullet TRUE for bullet style plot with mean and median inspired by Muller's work, or FALSE for violin plots.
     
jb_ploteq5d()}
\usage{
jb_ploteq5d(data, bullet = F)
}
\description{
bulleteq5d: Vertical Bullet Graph
 
 @author James Black - a ggplot take on the bullet plots by Simon Müller
 
 ""              ""         "" ""            "timepoints" 
 @param data:  data.frame that is produced by rbinding runs of jb_geteq5dvalues
                    bar  : number of bar values are from
                    eq_index     : eq5d index score
                    impaired_domains     : number of domains listed as impaired
                    
 @param bullet TRUE for bullet style plot with mean and median inspired by Muller's work, or FALSE for violin plots.
     
jb_ploteq5d()
}

