% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jb_pushoverr.R
\name{jb_pushiteration}
\alias{jb_pushiteration}
\title{Send push notifications to your phone from R for each n iterations of i}
\usage{
jb_pushiteration(message = "Script is on iteration: ",
  iteration_interval = 10)
}
\arguments{
\item{message}{What you want to be in the text of the message}

\item{iteration_interval}{Send message for every n iterations}
}
\description{
Token is a text file in working directory. Pushover account required.
}
\examples{
# Read in token
token <- readChar("token",
                 file.info("token")$size) # shrink to character length
# Load settings
pushoverr::set_pushover_app(
  token=token,
  user="umTsDu9PSNCSf92NXEY21jUsvz3faT")
for(i in 1:1000){
 # send a message every 100 iterations
 jb_pushiteration(iteration_interval=100)
 }
}
\keyword{iphone}

