% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jb_summaryby.R, R/jb_summaryfunctions.R
\name{jb_summaryby}
\alias{jb_summaryby}
\title{Automatically install missing packages}
\usage{
jb_summaryby(variable, by)

jb_summaryby(variable, by)
}
\arguments{
\item{variable}{Vector of interest.}

\item{by}{The name of the package.}

\item{name}{The name of the package.}
}
\description{
Packager loader that will automatically install missing packages

Packager loader that will automatically install missing packages
}
\examples{
data <- iris
jb_summaryby(Sepal.Width,Species)
data <- iris
jb_summaryby(data$Sepal.Width,data$Species)
}
\keyword{packages}
\keyword{summary}

