% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jb_vioplot.R
\name{jb_vioplot}
\alias{jb_vioplot}
\title{Violin plots}
\usage{
jb_vioplot(Var1_left, Var1_right, Var2_left, Var2_right, title = "Title",
  xlabel = "xlabel", group1title = "Group 1", group2title = "Group 2")
}
\arguments{
\item{Var1_left}{Variable one, left side.}

\item{Var1_right}{Variable one, right side.}

\item{Var2_left}{Variable two, left side.}

\item{Var2_right}{Variable two, right side.}
}
\description{
Make violin plots - based off this https://github.com/mbjoseph/comdis/blob/master/vioplot2.R
}
\examples{
data <- iris
jb_vioplot(
 Var1_left,
 Var1_right,
 Var2_left,
 Var2_right,
 title="This is a title",
 xlabel="Something",
 group1title="On the left",
 group2title="On the right")
}
\keyword{packages}

