% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/jb_CutIntoBins.R
\name{jb_CutIntoBins}
\alias{jb_CutIntoBins}
\title{Make sparklines from a variable}
\usage{
jb_CutIntoBins(variable, bins = 10, width = 6)
}
\arguments{
\item{bins}{Number of bins or bars in the histogram. Defaults to 10.}

\item{width}{The width of the histogram. Defaults to 6.}
}
\description{
Sparkline histograms form an R variable
}
\examples{
jb_CutIntoBins()
}
\keyword{latex}

