% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jb_eq5d.R
\name{jb_geteq5dvalues}
\alias{jb_geteq5dvalues}
\title{jb_geteq5dvalues: Get summary stats from EQ-5D
 
 @author James Black - a ggplot take on the bullet plots by Simon Müller
 
 @param barorder Numeric 1 to whatever. For which bar (or timepoint)  this set of EQ-5D measurements will be on the plot.
                    
 @param eq_index The main EQ-5D index score. 
     
 @param domain_one Any of the five eq domains - order doesn't matter
 
 @param domain_two Any of the five eq domains - order doesn't matter}
\usage{
jb_geteq5dvalues(barorder, eq_index, domain_one, domain_two, domain_three,
  domain_four, domain_five)
}
\description{
@param domain_three Any of the five eq domains - order doesn't matter
  
 @param domain_four Any of the five eq domains - order doesn't matter
}
\details{
@param domain_five Any of the five eq domains - order doesn't matter
   
 jb_geteq5dvalues()
}

