% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jb_eq5d.R
\name{jb_ploteq5d}
\alias{jb_ploteq5d}
\title{bulleteq5d: Vertical Bullet Graph}
\usage{
jb_ploteq5d(data, bullet = F)
}
\arguments{
\item{bullet}{TRUE for bullet style plot with mean and median inspired by Muller's work, or FALSE for violin plots.

jb_ploteq5d()}

\item{data:}{data.frame with
                    measure  : label of what's being measured
                    unit     : label of units of the measure
                    high     : the high value for the measure
                    second_cut     : the mean value for the measure
                    first_cut  : the first_cut value for the measure
                    target   : the target value for the measure
                    value    : the actual value of the measure}
}
\description{
bulleteq5d: Vertical Bullet Graph
}
\author{
James Black - a ggplot take on the bullet plots by Simon Müller
}

